+++ src/server/js/publish.js
@@ -2,11 +2,11 @@
  * pwix:env-settings/src/server/js/publish.js
  */
 
-// the ready() reactive function obviously doesn't automgically propagate to the clients
+// the ready() reactive function obviously doesn't automagically propagate to the clients
 // in order to have a ready() reactive function on the client-side, the server publishes the status, and the client subscribes to this publication
 
-Meteor.publish( EnvSettings.C._publicationName, function(){
-    this.added( EnvSettings.C._collectionName, '_ready', { ready: EnvSettings.ready() });
-    this.changed( EnvSettings.C._collectionName, '_ready', { ready: EnvSettings.ready() });
+Meteor.publish( EnvSettings.C.readyPub.publication, function(){
+    this.added( EnvSettings.C.readyPub.collection, '_ready', { ready: EnvSettings.ready() });
+    this.changed( EnvSettings.C.readyPub.collection, '_ready', { ready: EnvSettings.ready() });
     this.ready();
 });